Move main() function to common code
Summary:
    - Move BB and CL main.c to common layer
    - Project defined initial behavior is in project code "src/platform/plat_init.c"
